Formalize client vs server send marker behavior
Currently, all our clients set a send marker and none of our
servers do, meaning that only the clients send data by default.
This means that any connection ends when the client is done sending
data and gives processes a clear lifetime.

However, that behavior was very subtle, so I'm making it explicit.
I also added a provider method to make it easier to reuse the same
send marker for the server.