Fixed typo in __CBMC_havok -> __CMBC_havoc
Actually dropped __CBMC_havoc as properties no longer hold with it
becuase of the way it havoc whole objects.

Instead randomizing an arbitrary byte, which isn't really enough, but
should catch most issues.

Also dropped ensures clauses that the whole *conn struct is unchanged
sometimes because this slows CBMC down too much (like I've not been
prepared to wait for it to finish).