Add API to generate local host session ticket keys
The goal of the API(s2n_config_generate_and_add_ticket_crypto_key) is to
make it simpler for applications that want to generate a session ticket
key that will be used locally on the host(not shared among a fleet of
servers). The API takes no secrets as argument from the user to
prevent any possibility of misuse.

Before this API was added, applications generated a session local
session ticket by generating secret key material and name
themselves and then providing them as arguments to
s2n_config_add_ticket_crypto_key.