Fix s2n_self_talk_broken_pipe_test on OSX
Currently the s2n_self_talk_broken_pipe_test fails on OSX. After
investigation this was due to a race condition. After tweaking the
server sleep to be a bit longer for the client to close the IO pipe, the
test started passing.

In the future we should consider a solution that synchronizes the two
processes that is not reliant on sleep.

The result in the was using the wrong size type in the test, so I
updated that as well to align with the s2n_send API.