Fix MacOS unit tests (#3278)
Currently the s2n_self_talk_broken_pipe_test fails on MacOS. After
investigation this was due to a race condition. Using the FreeBSD code tweaks fixes this issue. See https://github.com/aws/s2n-tls/pull/3184 for more context.

The result in the was using the wrong size type in the test, so I updated that as well to align with the s2n_send API.

s2n_mem_test is not implemented for MacOS so I have disabled it to avoid creating errors.