Add buffered send integration test
Add buffered send option to s2nc and s2nd. Implement a more reliable send
buffer integration test. Enforce a minimum fragment size of 2 for alert
messages.

The send buffer size option for s2nc and s2nd doesn't included a short option,
perhaps new options in the future shouldn't include a short option.

With very large numbers integration tests, resulting from a large number of
different pytest parameters I observed failures that appeared to be related to
the test framework, and not the TLS implementation. Namely, a failure in which
both the client and server would continually timeout without ever making
contact with the other side. A solution to this is decreasing the number of
test cases, while maintaining coverage over the values of interest. In this
case that is 1) making sure the send buffer works with all different
protocols/ciphers and 2) making sure all different buffer sizes work. We also
wanted to make sure that the send buffer works on both the client and server
side.

There were a few separate edge case issues with the public API. First the
documentation was incorrect about the minimum possible value for the size of
the send buffer. Second the actual minimum size would result in an always
fragmented alert message. While that's not an issue for most protocols, there
is a known [attack](https://www.mitls.org/pages/attacks/Alert) against TLS 1.3
involving fragmented alert messages. GNU TLS chose to reject fragmented alerts
and as a result, when the send buffer is set too small, we fail to connect to
GNU TLS. This commit bumps that minimum up such that the minimum fragment size
is 2 (the length of an alert message).