Allow static and shared libs to be mixed.
Currently, users cannot build a shared lib that uses a static libs2n.a, and they cannot build a static lib that uses a shared libs2n.so.

With this change, a library can use either static libs2n.a or shared libs2n.so. If both are static and shared libs2n are installed, BUILD_SHARED_LIBS serves as a tie-breaker.