Fix time line send data - resolve flakiness
Previously these tests would inexplicably fail, and often need to be
automatically retried in order to PASS. This was a result of the time
line for sending and receiving data being unclear. By making the order
of events (including send times *and* close times) the tests no longer
mysteriously fail.

One particularly tricky aspect of this is that the data must be sent
after the handshake is complete. But not all cli utilities are equally
clear about when that moment is. S2N always prints "s2n is ready", but
GnuTLS is not so clear. The only tool the framework provides to find
this moment is a so-called send marker. That is a string to search for
on stdout after which data is to be provided to stdin. To find those
strings and ensure that the first instance of them was on stdout when
the handshake was ready took some time.

This strategy for making the tests more reliable also required that we
have a send marker on the server. In this case I used the connected
client to send a send marker to the server, which will appear on its
stdout.

Lastly I found that, due to a small send buffer and resulting message
fragmentation, GnuTLS and OpenSSL would often choose to print the
received app data in a non-contiguous fashion on stdout. So I needed
to resort to a subsequence (not substring) check to ensure the peer
received the application data.

Currently I am limiting the selection of certs only to ensure the test
takes a reasonable amount of time. In a test framework that has better
support for parallelism I would advise using them all, but the test is
already around ~15 minutes (not including the CI setup and teardown).