Avoid mysterious issue with long running tests
With very long running tests something goes wrong with the test that
results in the following output:

ERROR: InvocationError: '...'

I've experimented by increasing the `--log-cli-level` to INFO
and DEBUG. That results in slightly more legible output that indicates
which test failed -- however which test fails is different every time.
Whats more the failing result can't be replicated when the test is run
alone (or even as part of a subset).

I tried changing the number of threads to 1 (`-n 1`) however this just
takes longer and  achieved the same  inconsistent result.

I tried running three separate sessions runs, with each run only
including one of the three listed certs. All three runs passed.

I tried the same with the FRAGMENT_PREFERENCE. Using only one of the
three options on each of the three separate runs. All three of those
runs passed.

Lastly I tried reducing the value of SEND_DATA_SIZE to 2^15. This also
works and makes all of the tests pass.

These test runs generally last 15+ minutes and fail after 90% + of the
tests have run. I'm convinced that the test framework is failing and
not the individual tests, because all the tests pass on their own or
when subdivided into groups, but not all together.

I wish the test framework had the ability to give me more information,
without that I don't know what to do except try to avoid the problem.

This commit tries to maintain the maximum number of test combinations
while side stepping the issue.

Reducing the size of the data to send to 2^15 shouldn't affect the
efficacy of this test since we still have send buffer sizes that are
bigger and smaller than it (as we did when it was 2^16). The
application data total size is also still larger than the maximum data
allowed per record (2^14). The reduction does make the tests run
significantly faster and eliminates the error we see from the framework.