Ensure extended master secrets ext have no data (#3588)
RFC 7627, section 5.2:
This document defines a new TLS extension, "extended_master_secret"
(with extension type 0x0017), which is used to signal both client and
server to use the extended master secret computation. The
"extension_data" field of this extension is empty. Thus, the entire
encoding of the extension is 00 17 00 00 (in hexadecimal.)

We should therefore ensure that there is no data in the extension.