Fix very minor DeprecationWarning in integrationv2
I discovered this while working on an integrationv2 test. I got the
following warning:

    DeprecationWarning: invalid escape sequence

The invalid escape sequence here is '\d'. In Python 3.6 and above this
warning was added. The suggested fix here is to use a raw string or
properly escape the backslash. For regex source raw strings are the
better choice.