clang-format `tls/extensions` and enforce in ci (#3633)
Change clang format to use ColumnLimit: 0. This disables the column limit leaving it up to the programmer to decide when to line break. We did this because clang-format was choosing some odd places to break lines. Also we have some instances where we really want to be in control of the line breaks.