Implement Deref and DerefMut traits for PooledConnection
This change is to implement the Deref and DerefMut traits for the PooledConnection data type. This was done so that customers could leverage Rust smart dereferencing in order to make method calls implemented in the inner Connection directly to the PooledConnection instance without having to preface each Connection method call with as_ref() or as_mut().

This change also implements the PartialEq trait on the Connection type for testing purposes.