Add config to read until error or supplied buffer is full (#3690)
The code appeared to be designed to read multple TLS records, until the buffer was full, however a break statement at the end limited the logic to a single TLS record. This change introduces a config which allows changing of this behavior to read as much as possible in order to fill the supplied buffer.

In order to maintain legacy behavior, the config is defaulted to off.