Add llvm source code coverage
I think the link/compile options might be too broad. Need to only enable
them when LLVM is the compiler. Otherwise should probably just enable
the coverage flag. Or perhaps the coverage flag should always be
enabled.

The reports are not particularly nice looking. We should add an
additional to link the summary table to the line level coverage. I think
if we just shove all of the things into relative html paths that should
be do-able.