Remove unused extension functions
Extension functionality was previously refactored. This commit removes a
number of the functions that are now no longer used. Some of the old
extension functions are still being used in unit tests, and this commit
does not touch those. This commit only removes functionality that
currently has no execution in the unit test suite.