sketch out CI config for coverage
This commit adds a few files that will be necessary for code coverage in
CI runs. It breaks a little bit from tradition by avoiding the
omni-script, but I think that is quite justifiable in this case since
the unit test setup in the omniscript (s2n_codebuild.sh) doesn't work
for me anyways.

I will be attempting to slot most functionality into the correct
codebuild phases, since it will make the codebuild logs a bit nicer to
look at.