remove unused env variables and clean scripts
The last run was able to successfully go through my manually invoked CI
run. Moving towards pushing this into CR. I didn't find out that code
cov really doesn't like some of the branch misses in the macros's, and
our overall codebase coverage is significantly lower than codecov
because of that. There doesn't seem to be a way to turn it off either.