ci: remove flaky build-dashboard action
issue here: https://github.com/aws/s2n-tls/issues/3763. This action does
not run, since the issue-dashboard relies on node 12, which is
deprecated and no longer runs in GH actions.

This commit entirely removes the workflow. Once the team has a solution
this commit will likely be reverted and the action added back in, but
rather than just temporarily disabling it I think it's cleaner to just
entirely remove it while it's broken.