a less nuclear appraoch
I read the message wrong. The node deprecation thing is just a warning
and not an error. Thus we don't have to delete the whole thing. I still
hate the idea of note knowing whether an action is going to succeed
before I click merge though, so now going with this appraoch. Honestly,
I'd like to do the same thing for our documentation, but probably going
to get more pushback on that.