Reinstate Kyber KEM check
Some AWS-LC versions with API v20 (such as v1.3.0) do not have the new
KEM API present, so this test would erroneously fail against slightly
older versions of AWS-LC. Now that [AWS-LC has incremented its API][1]
version, we can reliably key off of the new version to ensure that the
KEM API really is present.

For non-AWS-LC or older AWS-LC without an API version, we don't compile
the check because we only expect the Kyber KEM API to be present in
newer AWS-LC builds.

This commit resolves [Issue #3902][2].

[1]: https://github.com/aws/aws-lc/pull/900
[2]: https://github.com/aws/s2n-tls/issues/3902