test: fix tokio binding close test
The shutdown_after_split() test in the tokio bindings was secretly
failing due to ignoring errors on the try_join. The reason it was
failing was because of new half-close behavior for TLS 1.3 connection.

This commit fixes the existing test by forcing it to use TLS 1.2, and
adds a new test for the half close behavior of TLS 1.3