Pin pytest-rerunfailures version in integration tests
We've been seeing many failures of the `test_well_known_endpoints`
integration tests in PR CI. Upon inspecting the output, I noticed that
there was none of the [expected "rerun" output][1] in our failed
codebuild jobs. Examining the pytest-rerunfailures plugin's changelog
indicated that [version 11.0 dropped support for pytest 5.x][2]. We
currently pin our pytest version to 5.3 and do not pin
pytest-rerunfailures, so tox takes the latest version of the
pytest-rerunfailures, which is incompatible with our version of pytest.

[1]: https://github.com/pytest-dev/pytest-rerunfailures/#output
[2]: https://github.com/pytest-dev/pytest-rerunfailures/blob/master/CHANGES.rst#110-2023-01-12
[3]: https://github.com/aws/s2n-tls/blob/2da0bf416cae49bc2a58222494c1b3ef5aa7f492/tests/integrationv2/tox.ini#L12