refactor: ossl verify cb logic
The default X509_STORE_CTX only allows for a single verify callback, but
s2n-tls sometimes requires multiple pieces of functionality for that
callback. Previously we did this by chaining function together, but this
will become more difficult as we add more verify functions that we wish
to call.

This commit adds a generic config and callback that allows for verify
callbacks to be more easily configured.