Actually fix build
The interned builds were failing because the binaries couldn't access
the libcrypto method to check the version. I tried conditional compilation,
but different versions of libcrypto alias the version method to different
methods and the check was getting too messy. So instead, I'm just going
to call an s2n-tls method. s2nc and s2nd shouldn't do that, but they
do that all over the place anyway.